refactor(daemon-runtime): split the preview runtime out of createProjectRuntime
createProjectRuntime interleaved two concerns in one ~360-line factory: spawning runs
into worktrees, and on-demand app previews. They share almost nothing -- only the
project resolver and the composite key scheme -- and the coupling runs one way: a
finished run stops its own preview (#797, in tearDownWorktree), while previews know
nothing about runs.

So the cut follows that seam rather than forcing a symmetric split. Move the preview
state (activePreviews, lastServeTarget), its tracker, and the four preview RPCs plus
dispose into createPreviewRuntime in preview-runtime.ts, injected with { homeId,
resolveProject }. createProjectRuntime composes it, exposes its handlers as `preview`,
and reaches in only at the one stop-on-teardown call. The `<projectKey>::<runId>` key
scheme both halves encode moves to runtime-keys.ts so it keeps one home across the two
modules (no import cycle). Behavior, RPC signatures, and the ProjectRuntime surface are
unchanged.
